Created separate files to work on assembler and splitted the emulator
We had to split the emulator into small files since it was massive. However, the use of structs made it difficult to do so. At the end we could only split the main file into two files: one with functions using structs and one without. To avoid this problem in the future we decided to split the files from the beginning.